Add Property Detail Screen, ExchangeRate API files

+++ app/src/main/java/com/dscoding/hosteluniverse/properties/data/model/LocationDto.kt
@@ -9,8 +9,8 @@ data class LocationDto(
 
 @Serializable
 data class CityDto(
-    val id: Long,
+    val id: Int,
     val name: String,
-    val idCountry: Long,
+    val idCountry: Int,
     val country: String
 )
