+++ app/src/main/java/com/dscoding/hosteluniverse/properties/domain/model/OverallRating.kt
@@ -1,6 +1,6 @@
 package com.dscoding.hosteluniverse.properties.domain.model
 
 data class OverallRating(
-        val overall: Long,
-        val numberOfRatings: String
+    val overall: Int,
+    val numberOfRatings: String
     )
